requirements.txt
tornado>=3.2
torndb>=0.2
PyYAML>=3.10
pycrypto>=2.6.1
MySQL-python>=1.2.5
hiredis>=0.1.3
redis>=2.10.1
17MonIP==0.2.0
ua-parser==0.3.5
user-agents==0.3.0
